疑似DBの削除
