+++ src/Backend/entities/problem.ts
@@ -1,0 +1,5 @@
+export type Problem = {
+  id: number;
+  title: string;
+  description: string;
+};
