+++ src/Backend/entities/submission.ts
@@ -1,0 +1,6 @@
+export type Submission = {
+  id: number;
+  problemId: number;
+  userName: string;
+  code: string;
+};
